Module Protocol.Voting_period_storage

val init : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> Tezos_raw_protocol_008_PtEdoTez.Voting_period_repr.t -> (Tezos_raw_protocol_008_PtEdoTez.Raw_context.t, Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.error list) Stdlib.result Lwt.t

val init_first_period : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> start_position:int32 -> (Tezos_raw_protocol_008_PtEdoTez.Raw_context.t, Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.error list) Stdlib.result Lwt.t

Sets the initial period to {voting_period = root; kind = Proposal; start_position}.

val reset : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> (Tezos_raw_protocol_008_PtEdoTez.Raw_context.t, Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.error list) Stdlib.result Lwt.t

Increment the index by one and set the kind to Proposal.

val succ : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> (Tezos_raw_protocol_008_PtEdoTez.Raw_context.t, Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.error list) Stdlib.result Lwt.t

Increment the index by one and set the kind to its successor.

val get_current : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> (Tezos_raw_protocol_008_PtEdoTez.Voting_period_repr.t, Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.error list) Stdlib.result Lwt.t

val get_current_kind : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> (Tezos_raw_protocol_008_PtEdoTez.Voting_period_repr.kind, Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.error list) Stdlib.result Lwt.t

val is_last_block : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> (bool, Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.error list) Stdlib.result Lwt.t

Returns true if the context level is the last of current voting period.

val get_current_info : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> (Tezos_raw_protocol_008_PtEdoTez.Voting_period_repr.info, Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.error list) Stdlib.result Lwt.t

val get_rpc_fixed_current_info : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> (Tezos_raw_protocol_008_PtEdoTez.Voting_period_repr.info, Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.error list) Stdlib.result Lwt.t

val get_rpc_fixed_succ_info : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> (Tezos_raw_protocol_008_PtEdoTez.Voting_period_repr.info, Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.error list) Stdlib.result Lwt.t