Module Tezos_raw_protocol_008_PtEdoTez.Apply_results

Result of applying an operation, can be used for experimenting with protocol updates, by clients to print out a summary of the operation at pre-injection simulation and at confirmation time, and by block explorers.

type 'kind operation_metadata = {

contents : 'kind contents_result_list;

}

Result of applying a Operation.t. Follows the same structure.

and packed_operation_metadata =

| Operation_metadata : 'kind operation_metadata -> packed_operation_metadata

| No_operation_metadata : packed_operation_metadata

and 'kind contents_result_list =

| Single_result : 'kind contents_result -> 'kind contents_result_list

| Cons_result : 'kind Alpha_context.Kind.manager contents_result * 'rest Alpha_context.Kind.manager contents_result_list -> ('kind * 'rest) Alpha_context.Kind.manager contents_result_list

Result of applying a Operation.contents_list. Follows the same structure.

and packed_contents_result_list =

| Contents_result_list : 'kind contents_result_list -> packed_contents_result_list

and 'kind contents_result =

| Endorsement_result : {

balance_updates : Alpha_context.Delegate.balance_updates;

delegate : Tezos_crypto.Signature.public_key_hash;

slots : int list;

} -> Alpha_context.Kind.endorsement contents_result

| Seed_nonce_revelation_result : Alpha_context.Delegate.balance_updates -> Alpha_context.Kind.seed_nonce_revelation contents_result

| Double_endorsement_evidence_result : Alpha_context.Delegate.balance_updates -> Alpha_context.Kind.double_endorsement_evidence contents_result

| Double_baking_evidence_result : Alpha_context.Delegate.balance_updates -> Alpha_context.Kind.double_baking_evidence contents_result

| Activate_account_result : Alpha_context.Delegate.balance_updates -> Alpha_context.Kind.activate_account contents_result

| Proposals_result : Alpha_context.Kind.proposals contents_result

| Ballot_result : Alpha_context.Kind.ballot contents_result

| Manager_operation_result : {

balance_updates : Alpha_context.Delegate.balance_updates;

operation_result : 'kind manager_operation_result;

internal_operation_results : packed_internal_operation_result list;

} -> 'kind Alpha_context.Kind.manager contents_result

Result of applying an Operation.contents. Follows the same structure.

and packed_contents_result =

| Contents_result : 'kind contents_result -> packed_contents_result

and 'kind manager_operation_result =

| Applied of 'kind successful_manager_operation_result

| Skipped : 'kind Alpha_context.Kind.manager -> 'kind manager_operation_result

The result of an operation in the queue. Skipped ones should always be at the tail, and after a single Failed.

and _ successful_manager_operation_result =

| Reveal_result : {

consumed_gas : Alpha_context.Gas.Arith.fp;

} -> Alpha_context.Kind.reveal successful_manager_operation_result

| Transaction_result : {

storage : Alpha_context.Script.expr option;

lazy_storage_diff : Alpha_context.Lazy_storage.diffs option;

balance_updates : Alpha_context.Delegate.balance_updates;

originated_contracts : Alpha_context.Contract.t list;

consumed_gas : Alpha_context.Gas.Arith.fp;

storage_size : Z.t;

paid_storage_size_diff : Z.t;

allocated_destination_contract : bool;

} -> Alpha_context.Kind.transaction successful_manager_operation_result

| Origination_result : {

lazy_storage_diff : Alpha_context.Lazy_storage.diffs option;

balance_updates : Alpha_context.Delegate.balance_updates;

originated_contracts : Alpha_context.Contract.t list;

consumed_gas : Alpha_context.Gas.Arith.fp;

storage_size : Z.t;

paid_storage_size_diff : Z.t;

} -> Alpha_context.Kind.origination successful_manager_operation_result

| Delegation_result : {

consumed_gas : Alpha_context.Gas.Arith.fp;

} -> Alpha_context.Kind.delegation successful_manager_operation_result

Result of applying a manager_operation_content, either internal or external.

and packed_successful_manager_operation_result =

| Successful_manager_result : 'kind successful_manager_operation_result -> packed_successful_manager_operation_result

and packed_internal_operation_result =

| Internal_operation_result : 'kind Alpha_context.internal_operation * 'kind manager_operation_result -> packed_internal_operation_result

val operation_metadata_encoding : packed_operation_metadata Tezos_base__TzPervasives.Data_encoding.t

Serializer for packed_operation_result.

val operation_data_and_metadata_encoding : (Alpha_context.Operation.packed_protocol_data * packed_operation_metadata) Tezos_base__TzPervasives.Data_encoding.t

type 'kind contents_and_result_list =

| Single_and_result : 'kind Alpha_context.contents * 'kind contents_result -> 'kind contents_and_result_list

| Cons_and_result : 'kind Alpha_context.Kind.manager Alpha_context.contents * 'kind Alpha_context.Kind.manager contents_result * 'rest Alpha_context.Kind.manager contents_and_result_list -> ('kind * 'rest) Alpha_context.Kind.manager contents_and_result_list

type packed_contents_and_result_list =

| Contents_and_result_list : 'kind contents_and_result_list -> packed_contents_and_result_list

val contents_and_result_list_encoding : packed_contents_and_result_list Tezos_base__TzPervasives.Data_encoding.t

val pack_contents_list : 'kind Alpha_context.contents_list -> 'kind contents_result_list -> 'kind contents_and_result_list

val unpack_contents_list : 'kind contents_and_result_list -> 'kind Alpha_context.contents_list * 'kind contents_result_list

val to_list : packed_contents_result_list -> packed_contents_result list

val of_list : packed_contents_result list -> packed_contents_result_list

type ('a, 'b) eq =

| Eq : ('a, 'a) eq

val kind_equal_list : 'kind Alpha_context.contents_list -> 'kind2 contents_result_list -> ('kind, 'kind2) eq option

type block_metadata = {

baker : Tezos_crypto.Signature.public_key_hash;

level : Alpha_context.Level.compat_t;

level_info : Alpha_context.Level.t;

voting_period_kind : Alpha_context.Voting_period.kind;

voting_period_info : Alpha_context.Voting_period.info;

nonce_hash : Nonce_hash.t option;

consumed_gas : Alpha_context.Gas.Arith.fp;

deactivated : Tezos_crypto.Signature.public_key_hash list;

balance_updates : Alpha_context.Delegate.balance_updates;

}

val block_metadata_encoding : block_metadata Tezos_base__TzPervasives.Data_encoding.t