Module Delegate_services.Baking_rights

type t = {

level : Alpha_context.Raw_level.t;

delegate : Tezos_crypto.Signature.public_key_hash;

priority : int;

timestamp : Alpha_context.Timestamp.t option;

}

val get : 'a Tezos_protocol_environment_008_PtEdoTez__Environment.RPC_context.#simple -> ?levels:Alpha_context.Raw_level.t list -> ?cycles:Alpha_context.Cycle.t list -> ?delegates:Tezos_crypto.Signature.public_key_hash list -> ?all:bool -> ?max_priority:int -> 'a -> (t list, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

Retrieves the list of delegates allowed to bake a block.

By default, it gives the best baking priorities for bakers that have at least one opportunity below the 64th priority for the next block.

Parameters levels and cycles can be used to specify the (valid) level(s) in the past or future at which the baking rights have to be returned. Parameter delegates can be used to restrict the results to the given delegates. If parameter all is true, all the baking opportunities for each baker at each level are returned, instead of just the first one.

Returns the list of baking slots. Also returns the minimal timestamps that correspond to these slots. The timestamps are omitted for levels in the past, and are only estimates for levels later that the next block, based on the hypothesis that all predecessor blocks were baked at the first priority.