Module type Script_interpreter.STEP_LOGGER

STEP_LOGGER is the module type of logging modules as passed to the Michelson interpreter. Note that logging must be performed by side-effects on an underlying log structure.

val log_interp : Alpha_context.context -> ('bef, 'aft) Script_typed_ir.descr -> 'bef -> unit

log_interp is called at each call of the internal function interp. interp is called when starting the interpretation of a script and subsequently at each Exec instruction.

val log_entry : Alpha_context.context -> ('bef, 'aft) Script_typed_ir.descr -> 'bef -> unit

log_entry is called before executing each instruction but after gas for this instruction has been successfully consumed.

val log_exit : Alpha_context.context -> ('bef, 'aft) Script_typed_ir.descr -> 'aft -> unit

log_exit is called after executing each instruction.

val get_log : unit -> (execution_trace option, Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.error list) Stdlib.result Lwt.t

get_log allows to obtain an execution trace, if any was produced.