Module Tezos_raw_protocol_008_PtEdoTez__Roll_storage

Basic roll manipulation.

If storage related to roll (i.e. `Storage.Roll`) is not used outside of this module, this interface enforces the invariant that a roll is always either in the limbo list or in a contract list.

type Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.error += Consume_roll_change | No_roll_for_delegate | No_roll_snapshot_for_cycle of Tezos_raw_protocol_008_PtEdoTez.Cycle_repr.t | Unregistered_delegate of Tezos_protocol_environment_008_PtEdoTez__Environment.Signature.Public_key_hash.t

val init : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> Tezos_raw_protocol_008_PtEdoTez.Raw_context.t Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.tzresult Tezos_protocol_environment_008_PtEdoTez__Environment.Lwt.t

val init_first_cycles : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> Tezos_raw_protocol_008_PtEdoTez.Raw_context.t Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.tzresult Tezos_protocol_environment_008_PtEdoTez__Environment.Lwt.t

val cycle_end : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> Tezos_raw_protocol_008_PtEdoTez.Cycle_repr.t -> Tezos_raw_protocol_008_PtEdoTez.Raw_context.t Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.tzresult Tezos_protocol_environment_008_PtEdoTez__Environment.Lwt.t

val snapshot_rolls : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> Tezos_raw_protocol_008_PtEdoTez.Raw_context.t Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.tzresult Tezos_protocol_environment_008_PtEdoTez__Environment.Lwt.t

val fold : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> f:(Tezos_raw_protocol_008_PtEdoTez.Roll_repr.roll -> Tezos_protocol_environment_008_PtEdoTez__Environment.Signature.Public_key.t -> 'a -> 'a Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.tzresult Tezos_protocol_environment_008_PtEdoTez__Environment.Lwt.t) -> 'a -> 'a Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.tzresult Tezos_protocol_environment_008_PtEdoTez__Environment.Lwt.t

val baking_rights_owner : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> Tezos_raw_protocol_008_PtEdoTez.Level_repr.t -> priority:int -> Tezos_protocol_environment_008_PtEdoTez__Environment.Signature.Public_key.t Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.tzresult Tezos_protocol_environment_008_PtEdoTez__Environment.Lwt.t

val endorsement_rights_owner : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> Tezos_raw_protocol_008_PtEdoTez.Level_repr.t -> slot:int -> Tezos_protocol_environment_008_PtEdoTez__Environment.Signature.Public_key.t Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.tzresult Tezos_protocol_environment_008_PtEdoTez__Environment.Lwt.t

module Delegate : sig ... end

module Contract : sig ... end

val delegate_pubkey : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> Tezos_protocol_environment_008_PtEdoTez__Environment.Signature.Public_key_hash.t -> Tezos_protocol_environment_008_PtEdoTez__Environment.Signature.Public_key.t Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.tzresult Tezos_protocol_environment_008_PtEdoTez__Environment.Lwt.t

val get_rolls : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> Tezos_protocol_environment_008_PtEdoTez__Environment.Signature.Public_key_hash.t -> Tezos_raw_protocol_008_PtEdoTez.Roll_repr.t list Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.tzresult Tezos_protocol_environment_008_PtEdoTez__Environment.Lwt.t

val get_change : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> Tezos_protocol_environment_008_PtEdoTez__Environment.Signature.Public_key_hash.t -> Tezos_raw_protocol_008_PtEdoTez.Tez_repr.t Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.tzresult Tezos_protocol_environment_008_PtEdoTez__Environment.Lwt.t

val update_tokens_per_roll : Tezos_raw_protocol_008_PtEdoTez.Raw_context.t -> Tezos_raw_protocol_008_PtEdoTez.Tez_repr.t -> Tezos_raw_protocol_008_PtEdoTez.Raw_context.t Tezos_protocol_environment_008_PtEdoTez__Environment.Error_monad.tzresult Tezos_protocol_environment_008_PtEdoTez__Environment.Lwt.t